Change func attr __args__ format
Before we embed argument's default value in __args__, such as:

[null, null, ['a'], ['b', 0]]

Actually the default value is never used. The design purpose of embed
default value is: python's default value is an variable slot, not value.
For example:

  def foo(b, a=[]):
		a.append(b)
		return a

The argument `a' default variable is changed after calling `foo(1)'.

This feature is not easy to implement without hurting performance. Such
as:

  $m.foo = function foo(b) {
		(a!==undefined) || a = $m.foo.__args__[3][1];
	}

	1. Resolve `$m.foo.__args__' is slow
	2. For methods, the function is more complex, `$m.C.foo.__args__[3][1]'

This feature is actually a side-effect of python implementation, and is
not wise to use a mutable value as default argument value. So jscompiler
drop this feature entirely, to reduce memory consume and increase
keyword calling (like `foo(a=1)') speed.